polycubed: move utils to new utils folder
The root folder of polycubed is getting messy, there are many files around
and it could get worse because were are planning to add some new files.

This commit creates a new utils folder and moves some files there.
